Move stuff to pdaaal::internal + use " " for #include
