+++ src/include/pdaaal/parsing/NfaParser.cpp
@@ -24,8 +24,8 @@
  * Created on 05-07-2021.
  */
 
-#include <pdaaal/parsing/NfaParser.h>
-#include <pdaaal/parsing/NfaParserGrammar.h>
+#include "NfaParser.h"
+#include "NfaParserGrammar.h"
 
 namespace pdaaal {
